add plane fix canvas

+++ src/vector_math.c
@@ -6,20 +6,20 @@
 /*   By: bdrinkin <bdrinkin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/21 20:38:12 by bdrinkin          #+#    #+#             */
-/*   Updated: 2020/10/22 04:11:07 by bdrinkin         ###   ########.fr       */
+/*   Updated: 2020/10/23 22:26:02 by bdrinkin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "rtv1.h"
 
-float		mod3(t_vec3 vec)
+double		mod3(t_vec3 vec)
 {
 	return (sqrt(vec.x * vec.x +
 				vec.y * vec.y +
 				vec.z * vec.z));
 }
 
-float		clamp(float x, float min, float max)
+double		clamp(double x, double min, double max)
 {
 	if (x > min && x < max)
 		return (x);
@@ -60,7 +60,7 @@ t_vec3		cross(t_vec3 a, t_vec3 b)
 	return result;
 }
 
-t_vec3		cross_scalar(t_vec3 vect, float scalar)
+t_vec3		cross_scalar(t_vec3 vect, double scalar)
 {
 	t_vec3	result;
 
@@ -73,7 +73,7 @@ t_vec3		cross_scalar(t_vec3 vect, float scalar)
 t_vec3		normalize(t_vec3 vector)
 {
 	t_vec3	v1;
-	float	lengt;
+	double	lengt;
 
 	lengt = mod3(vector);
 	v1.x = vector.x / lengt;
@@ -82,14 +82,14 @@ t_vec3		normalize(t_vec3 vector)
 	return (v1);
 }
 
-float		dot3(t_vec3 v1, t_vec3 v2)
+double		dot3(t_vec3 v1, t_vec3 v2)
 {
 	return (v1.x * v2.x +
 			v1.y * v2.y +
 			v1.z * v2.z);
 }
 
-float		acos_vec3(t_vec3 v1, t_vec3 v2)
+double		acos_vec3(t_vec3 v1, t_vec3 v2)
 {
 	return (acos(dot3(normalize(v1), normalize(v2))));
 }
